+++ src/test/java/blackjack/domain/profit/ResultTest.java
@@ -1,12 +1,11 @@
-package blackjack.domain.game;
+package blackjack.domain.profit;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.participant.Hand;
 import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Hand;
 import blackjack.domain.participant.Player;
 import blackjack.domain.participant.PlayerName;
-import blackjack.domain.profit.Result;
 import fixture.CardFixture;
 import fixture.HandFixture;
 import org.junit.jupiter.api.DisplayName;
